refactor: 도메인 수정
- user -> auth
