+++ src/main/java/com/example/webnovel/auth/domain/type/Role.java
@@ -1,0 +1,6 @@
+package com.example.webnovel.auth.domain.type;
+
+public enum Role {
+    USER,
+    ADMIN
+}
