+++ src/main/java/com/example/webnovel/auth/domain/type/UserType.java
@@ -1,0 +1,5 @@
+package com.example.webnovel.auth.domain.type;
+
+public enum UserType {
+    GENERAL
+}
