--- src/main/java/com/example/webnovel/user/domain/type/Role.java
@@ -1,6 +1,0 @@
-package com.example.webnovel.user.domain.type;
-
-public enum Role {
-    USER,
-    ADMIN
-}
