--- src/main/java/com/example/webnovel/user/domain/type/UserType.java
@@ -1,5 +1,0 @@
-package com.example.webnovel.user.domain.type;
-
-public enum UserType {
-    GENERAL
-}
